requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9
langchain==0.0.350
langchain-community==0.0.10
langchain-openai==0.0.2
chromadb==0.4.18
sentence-transformers==2.2.2
pypdf2==3.0.1
python-docx==1.1.0
python-multipart==0.0.6
jinja2==3.1.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
redis==5.0.1
celery==5.3.4
pandas==2.1.4
numpy==1.24.4
scikit-learn==1.3.2
streamlit==1.28.2
plotly==5.17.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
